requirements.txt
streamlit>=1.31.0
supabase>=2.3.4
python-dotenv>=1.0.0
pandas>=2.2.0
psycopg2-binary
python-dotenv